requirements.txt
django >= 2.2
simplejson
pyotp
python-u2flib-server
ua-parser
user-agents
python-jose
fido2 > 1.1.0,<2.0
jsonLookup
raven
pyre-check
